+++ test/helpers/test-helper.ts
@@ -8,11 +8,11 @@ import { Revision } from '../../src/entities/revision';
 import { FileImport } from '../../src/entities/file-import';
 import { CsvInfo } from '../../src/entities/csv-info';
 import { Source } from '../../src/entities/source';
-import { SourceType } from '../../src/enums/source-type.enum';
-import { DimensionType } from '../../src/enums/dimension-type.enum';
-import { SourceAction } from '../../src/enums/source-action.enum';
-import { DataLocation } from '../../src/enums/data-location.enum';
-import { ImportType } from '../../src/enums/import-type.enum';
+import { SourceType } from '../../src/enums/source-type';
+import { DimensionType } from '../../src/enums/dimension-type';
+import { SourceAction } from '../../src/enums/source-action';
+import { DataLocation } from '../../src/enums/data-location';
+import { ImportType } from '../../src/enums/import-type';
 import { Dimension } from '../../src/entities/dimension';
 import { DimensionInfo } from '../../src/entities/dimension-info';
 import { User } from '../../src/entities/user';
