+++ test/metadata-routes.test.ts
@@ -16,7 +16,7 @@ import { DimensionDTO } from '../src/dtos/dimension-dto';
 import { RevisionDTO } from '../src/dtos/revision-dto';
 import { FileImportDTO } from '../src/dtos/file-import-dto';
 import DatabaseManager from '../src/db/database-manager';
-import { DataLocation } from '../src/enums/data-location.enum';
+import { DataLocation } from '../src/enums/data-location';
 
 import { createFullDataset } from './helpers/test-helper';
 import { getTestUser } from './helpers/get-user';
